+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-# Changed
+##
+## Changed
 - fixed documents created with tinymce causing a block validation error
 - fixed inNewTab and showOpenButton attributes being ignored due to missing boolean casting.
+- changed the JWT secret key field to a password input with a show/hide toggle
 
 ## 2.3.0
 ## Changed